requirements.txt
aiohttp==3.7.4.post0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.0
discord.py==1.7.3
discord-py-interactions==3.0.2
discord-ui==5.1.6
idna==3.4
multidict==6.0.4
requests==2.31.0
requests-oauthlib==1.3.1
yarl==1.9.2